Improving GUI and double -> int

+++ src/robots/Transmitter.java
@@ -1,27 +1,27 @@
 package robots;
 
 public class Transmitter {
-    private double x;
+    private int x;
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
 
-    private double y;
+    private int y;
 
-    Transmitter(double x, double y) {
+    Transmitter(int x, int y) {
         this.x = x;
         this.y = y;
     }
